Module air_parser::ast

Structs

Ap

(ap argument result)

ApMap

(ap key value %map)

Call

(call (peer part of a triplet: PeerPart) (function part of a triplet: FunctionPart) [arguments] output)

Canon

(canon peer_id $stream #canon_stream)

CanonMap

(canon peer_id %stream_map #%canon_stream_map)

CanonStream

A canonicalized stream without a lambda.

CanonStreamMap

A canonicalized stream map without a lambda.

CanonStreamMapScalar

(canon peer_id %stream_map scalar)

CanonStreamMapWithLambda

A canonicalized stream map with a lambda.

CanonStreamWithLambda

A canonicalized stream with a lambda.

FoldScalar

(fold scalar_iterable iterator instruction)

FoldStream

(fold stream_iterable iterator instruction)

FoldStreamMap

(fold stream_map_iterable iterator instruction)

InstructionErrorAST

An error wrapper with an optional lens.

Match

(match left_value right_value instruction)

MisMatch

(mismatch left_value right_value instruction)

Never

(never)

New

(new variable instruction)

Next

(fold stream_iterable iterator instruction)

Null

(null)

Par

(par instruction instruction)

Scalar

A scalar value without a lambda.

ScalarWithLambda

A scalar value with a lambda expression.

Seq

(seq instruction instruction)

Span

Stream

A stream without a lambda.

StreamMap

A map based on top of a stream.

Triplet

Triplet represents a location of the executable code in the network. It is build from PeerPart and FunctionPart of a Call instruction.

Xor

(xor instruction instruction)

Enums

ApArgument

ApResult

CallOutputValue

Fail

(fail 1337 “error message”) (fail %last_error%) (fail value)

FoldScalarIterable

ImmutableValue

Represents all immutable values that is possible to set in AIR scripts.

ImmutableVariable

A variable that could be either scalar or stream without lambda.

ImmutableVariableWithLambda

A variable that could be either scalar or stream with possible lambda expression.

Instruction

NewArgument

Number

ResolvableToPeerIdVariable

Contains all variable variants that could be resolved to a peer id.

ResolvableToStringVariable

Contains all variable variants that could be resolved to a string type.

StreamMapKeyClause

Traits

PeerIDErrorLogable